[Tests] Don't rely on examples for plugin system tests
The PythonPluginSystemManagerImplementationFactory tests were relying
on examples in the `resources/examples` folder. This isn't great as they
are subject to change, and it breaks compartmentalisation.

This switches the PPMIF tests to use the same local stub plugins that
the PythonPluginSystem tests use. This means we dont have to worry about
things breaking when we migrate SAM to be an template manager, etc...

Part of #742

Signed-off-by: Tom Cowland <tom@foundry.com>