[CI] Add TODO to enable Pylint for _openassetio
Pylint has rudimentary support for parsing C extension modules, but
this support does not extend to submodules created within the C
extension module, which we make heavy use of.

So for now just add the required config commented out with a TODO to
revist later, once the issue is resolved upstream.

Signed-off-by: David Feltell <david.feltell@foundry.com>